Extract common api options
